fix(scripts): run lex codegen via a Node runner, not a shell pipe
`set -o pipefail` is NOT portable: GitHub's Ubuntu `sh` (dash) rejects it
(`set: Illegal option -o pipefail`), which broke CI. And no `xargs` option
replaces it — `-r` only skips empty input, it does not propagate a producer's
non-zero exit, so a crashed list-generator still yields a "successful" pipeline.

Replace the pipe entirely with scripts/run-lex-codegen.js: a single Node
process that builds the permission-set-excluded file list and spawns
`lex <subcommand> <args> <files…>` with the list as argv, forwarding lex's exit
code. One authoritative exit code, no shell builtins, no ARG_MAX (argv, not a
command line), no pipefail. Empty list or spawn failure → non-zero exit.

(Renamed from codegen-lexicon-files.js, which only printed the list.)
