+++ package.json
@@ -42,12 +42,12 @@
     "check": "npm run gen-api && npm run lint && npm run typecheck && npm run build && npm run test",
     "build": "rollup -c && npm run build:types",
     "build:types": "tsc --project tsconfig.build.json",
-    "//permissions": "permission-set lexicons are published as-is but have no TS shape; lex gen-* cannot codegen them. scripts/codegen-lexicon-files.js emits the NUL-delimited lexicon list to codegen, excluding permission-sets by content (single source of truth shared with generate-exports.js). Piped into `xargs -0 -r` under `set -o pipefail`: pipefail propagates a producer failure (xargs alone cannot — it would run lex on partial/empty input and report success); -r skips lex if the list is somehow empty; xargs avoids ARG_MAX.",
-    "gen-api": "set -o pipefail; node ./scripts/codegen-lexicon-files.js | xargs -0 -r lex gen-api --yes ./generated && npm run gen-index",
-    "gen-md": "set -o pipefail; node ./scripts/codegen-lexicon-files.js | xargs -0 -r lex gen-md --yes ./lexicons.md",
+    "//permissions": "permission-set lexicons are published as-is but have no TS shape; lex gen-* cannot codegen them. scripts/run-lex-codegen.js runs a lex subcommand over all lexicons EXCEPT permission-sets (excluded by content — the same check generate-exports.js uses). It is a single Node process that spawns lex with the file list as argv and forwards lex's exit code, so a producer failure propagates without `set -o pipefail` (not portable across npm's sh) and there is no ARG_MAX limit.",
+    "gen-api": "node ./scripts/run-lex-codegen.js gen-api --yes ./generated && npm run gen-index",
+    "gen-md": "node ./scripts/run-lex-codegen.js gen-md --yes ./lexicons.md",
     "gen-schemas-md": "node ./scripts/generate-schemas.js",
     "//gen-ts": "UNUSED - use gen-api instead",
-    "gen-ts": "set -o pipefail; node ./scripts/codegen-lexicon-files.js | xargs -0 -r lex gen-ts-obj > generated/DO-NOT-USE-lexicons.ts",
+    "gen-ts": "node ./scripts/run-lex-codegen.js gen-ts-obj > generated/DO-NOT-USE-lexicons.ts",
     "gen-index": "node ./scripts/generate-exports.js",
     "lex": "lex",
     "test": "vitest run",
